+++ config/config.default.js
@@ -5,7 +5,7 @@ module.exports = app => {
 
   exports.knex = {
     default: {
-      dialect: 'mysql',
+      dialect: 'mysql2',
       connection: {
         database: null,
       },