+++ test/fixtures/apps/mysqlapp-connectstring/config/config.js
@@ -4,8 +4,8 @@ exports.keys = "i7oBLQc43aiDCC9zxYrYNw==";
 
 exports.knex = {
   client: {
-    dialect: "mysql",
-    connection: "mysql://root:@localhost/test"
+    dialect: "mysql2",
+    connection: "mysql://root:longbridge@localhost/test"
   },
   agent: true
 };