+++ test/fixtures/apps/mysqlapp-disable/config/config.js
@@ -4,12 +4,13 @@ exports.keys = 'i7oBLQc43aiDCC9zxYrYNw==';
 
 exports.knex = {
   client: {
-    dialect: 'mysql',
+    dialect: 'mysql2',
     connection: {
       host: '127.0.0.1',
       port: 3306,
       user: 'root',
       database: 'test',
+      password: 'longbridge'
     },
   },
   app: false,