+++ test/fixtures/apps/mysqlapp-dynamic/config/config.js
@@ -7,11 +7,12 @@ exports.knex = {
 };
 
 exports.mysql1 = {
-  dialect: 'mysql',
+  dialect: 'mysql2',
   connection: {
     host: '127.0.0.1',
     port: 3306,
     user: 'root',
     database: 'test',
+    password: 'longbridge'
   },
 };