+++ test/fixtures/apps/mysqlapp-multi-client-new/config/config.js
@@ -5,30 +5,33 @@ exports.keys = 'i7oBLQc43aiDCC9zxYrYNw==';
 exports.knex = {
   clients: {
     db1: {
-      dialect: 'mysql',
+      dialect: 'mysql2',
       connection: {
         host: '127.0.0.1',
         port: 3306,
         user: 'root',
         database: 'test',
+        password: 'longbridge'
       },
     },
     db2: {
-      dialect: 'mysql',
+      dialect: 'mysql2',
       connection: {
         host: '127.0.0.1',
         port: 3306,
         user: 'root',
         database: 'test',
+        password: 'longbridge'
       },
     },
     db3: {
-      dialect: 'mysql',
+      dialect: 'mysql2',
       connection: {
         host: '127.0.0.1',
         port: 3306,
         user: 'root',
         database: 'test',
+        password: 'longbridge'
       },
     },
   },