+++ test/fixtures/apps/mysqlapp/config/config.js
@@ -5,12 +5,13 @@ exports.keys = 'i7oBLQc43aiDCC9zxYrYNw==';
 exports.knex = {
   client: {
     debug: true,
-    dialect: 'mysql',
+    dialect: 'mysql2',
     connection: {
       host: '127.0.0.1',
       port: 3306,
       user: 'root',
       database: 'test',
+      password: 'longbridge'
     },
   },
   agent: true